+++ rr_hw_interface/package.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>rr_hw_interface</name>
+  <version>0.0.1</version>
+  <description>
+   Suction pump interface. Supports schmaltz FXP pump.
+  </description>
+  <license>Commercial</license>
+  <maintainer email="gautham.manoharan@rapyuta-robotics.com">Gautham Manoharan</maintainer>
+  <maintainer email="bhuvanchandra.dv@rapyuta-robotics.com">Bhuvanchandra</maintainer>
+  <maintainer email="yu.okamoto@rapyuta-robotics.com">Yu Okamoto</maintainer>
+  
+  <buildtool_depend>catkin</buildtool_depend>
+  
+  <depend>roscpp</depend>
+
+  <exec_depend>message_generation</exec_depend>
+</package>
