+++ rr_suction_pump/include/rr_suction_pump/suction_pump.hpp
@@ -9,8 +9,8 @@
 #include <libsoc_board.h>
 #include <libsoc_gpio.h>
 
-#include <hw_interface/hw_interface.hpp>
-#include <hw_interface/gpio.hpp>
+#include <rr_hw_interface/hw_interface.hpp>
+#include <rr_hw_interface/gpio/libsoc_gpio.hpp>
 
 namespace rapyuta
 {
